Test iterates through all EPANET versions
