+++ packages/epanet-js-engine/package.json
@@ -10,7 +10,7 @@
   ],
   "scripts": {
     "build": "rollup -c && cp src/index.d.ts dist/index.d.ts && cp -r lib/. dist/",
-    "test": "node test/test.js",
+    "test": "node test/smoke-test.js",
     "prepublishOnly": "npm run build"
   },
   "exports": {
