+++ packages/epanet-js-engine/test/smoke-test.js
@@ -1,0 +1,57 @@
+import assert from 'assert'
+import { readFileSync } from 'fs'
+import { fileURLToPath } from 'url'
+import { dirname, join } from 'path'
+
+import { loader } from '../dist/index.js'
+
+const __dirname = dirname(fileURLToPath(import.meta.url))
+
+const VERSIONS = [
+  'v2.2',
+  'v2.3',
+  'v2.3.1',
+  'v2.3.2',
+  'v2.3.3',
+  'v2.3.4',
+  'v2.3.5',
+  'master',
+  'dev',
+]
+
+for (const version of VERSIONS) {
+  process.stdout.write(`[${version}] `)
+
+  const { default: factory } = await import(`../dist/${version}/EpanetEngine.js`)
+  const wasmBinary = readFileSync(join(__dirname, `../dist/${version}/EpanetEngine.wasm`))
+  const f = await loader(() => factory({ wasmBinary }))
+
+  f.FS.writeFile('net.inp', readFileSync(join(__dirname, 'network.inp')))
+
+  const phPtr = f.malloc(4)
+  f.EN_createproject(phPtr)
+  const ph = f.getValue(phPtr, 'i32')
+  f.free(phPtr)
+
+  const inpPtr = f.allocateUTF8('net.inp')
+  const rptPtr = f.allocateUTF8('net.rpt')
+  const outPtr = f.allocateUTF8('net.out')
+  f.EN_open(ph, inpPtr, rptPtr, outPtr)
+  f.free(inpPtr)
+  f.free(rptPtr)
+  f.free(outPtr)
+
+  f.EN_solveH(ph)
+  f.EN_solveQ(ph)
+  f.EN_report(ph)
+  f.EN_deleteproject(ph)
+
+  const rpt = f.FS.readFile('net.rpt', { encoding: 'utf8' })
+  const out = f.FS.readFile('net.out')
+
+  assert.ok(rpt.length > 0, 'net.rpt must not be empty')
+  assert.ok(out.length > 0, 'net.out must not be empty')
+
+  console.log('OK! EPANET report:')
+  console.log(rpt)
+}
